+++ modules/meta.analysis/R/meta.analysis.R
@@ -109,11 +109,10 @@ pecan.ma <- function(trait.data, prior.distns,
     ## check for excess missing data
 
     if (all(is.na(data[["obs.prec"]]))) {
-      if (verbose) {
-        writeLines("NO ERROR STATS PROVIDED, DROPPING RANDOM EFFECTS")
-      }
-      data$site <- rep(1, nrow(data))
-      data$trt  <- rep(0, nrow(data))
+      PEcAn.logger::logger.warn("NO ERROR STATS PROVIDED\n Check meta-analysis Model Convergence", 
+                  "and consider turning off Random Effects by", 
+                  "setting <random.effects>FALSE</random.effects>",
+                  "in your pecan.xml settings file ")
     }
 
     if (!random) {